+++ packages/pocket-ic/src/common/rest.rs
@@ -723,3 +723,135 @@ impl Topology {
             .map(|(id, _)| *id)
     }
 }
+
+#[derive(
+    Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, JsonSchema,
+)]
+pub enum CanisterHttpMethod {
+    GET,
+    POST,
+    HEAD,
+}
+
+#[derive(
+    Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, JsonSchema,
+)]
+pub struct CanisterHttpHeader {
+    pub name: String,
+    pub value: String,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, JsonSchema)]
+pub struct RawCanisterHttpRequest {
+    pub subnet_id: RawSubnetId,
+    pub request_id: u64,
+    pub http_method: CanisterHttpMethod,
+    pub url: String,
+    pub headers: Vec<CanisterHttpHeader>,
+    pub body: Option<Vec<u8>>,
+    pub max_response_bytes: Option<u64>,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
+pub struct CanisterHttpRequest {
+    pub subnet_id: Principal,
+    pub request_id: u64,
+    pub http_method: CanisterHttpMethod,
+    pub url: String,
+    pub headers: Vec<CanisterHttpHeader>,
+    pub body: Option<Vec<u8>>,
+    pub max_response_bytes: Option<u64>,
+}
+
+impl From<RawCanisterHttpRequest> for CanisterHttpRequest {
+    fn from(raw_canister_http_request: RawCanisterHttpRequest) -> Self {
+        Self {
+            subnet_id: candid::Principal::from_slice(
+                &raw_canister_http_request.subnet_id.subnet_id,
+            ),
+            request_id: raw_canister_http_request.request_id,
+            http_method: raw_canister_http_request.http_method,
+            url: raw_canister_http_request.url,
+            headers: raw_canister_http_request.headers,
+            body: raw_canister_http_request.body,
+            max_response_bytes: raw_canister_http_request.max_response_bytes,
+        }
+    }
+}
+
+impl From<CanisterHttpRequest> for RawCanisterHttpRequest {
+    fn from(canister_http_request: CanisterHttpRequest) -> Self {
+        Self {
+            subnet_id: canister_http_request.subnet_id.into(),
+            request_id: canister_http_request.request_id,
+            http_method: canister_http_request.http_method,
+            url: canister_http_request.url,
+            headers: canister_http_request.headers,
+            body: canister_http_request.body,
+            max_response_bytes: canister_http_request.max_response_bytes,
+        }
+    }
+}
+
+#[derive(
+    Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, JsonSchema,
+)]
+pub struct CanisterHttpReply {
+    pub status: u16,
+    pub headers: Vec<CanisterHttpHeader>,
+    pub body: Vec<u8>,
+}
+
+#[derive(
+    Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, JsonSchema,
+)]
+pub struct CanisterHttpReject {
+    pub reject_code: u64,
+    pub message: String,
+}
+
+#[derive(
+    Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, JsonSchema,
+)]
+pub enum CanisterHttpResponse {
+    CanisterHttpReply(CanisterHttpReply),
+    CanisterHttpReject(CanisterHttpReject),
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, JsonSchema)]
+pub struct RawMockCanisterHttpResponse {
+    pub subnet_id: RawSubnetId,
+    pub request_id: u64,
+    pub response: CanisterHttpResponse,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
+pub struct MockCanisterHttpResponse {
+    pub subnet_id: Principal,
+    pub request_id: u64,
+    pub response: CanisterHttpResponse,
+}
+
+impl From<RawMockCanisterHttpResponse> for MockCanisterHttpResponse {
+    fn from(raw_mock_canister_http_response: RawMockCanisterHttpResponse) -> Self {
+        Self {
+            subnet_id: candid::Principal::from_slice(
+                &raw_mock_canister_http_response.subnet_id.subnet_id,
+            ),
+            request_id: raw_mock_canister_http_response.request_id,
+            response: raw_mock_canister_http_response.response,
+        }
+    }
+}
+
+impl From<MockCanisterHttpResponse> for RawMockCanisterHttpResponse {
+    fn from(mock_canister_http_response: MockCanisterHttpResponse) -> Self {
+        Self {
+            subnet_id: RawSubnetId {
+                subnet_id: mock_canister_http_response.subnet_id.as_slice().to_vec(),
+            },
+            request_id: mock_canister_http_response.request_id,
+            response: mock_canister_http_response.response,
+        }
+    }
+}